Raise Python error from SerializeToString() for missing required fields.
Improving the debugging capabilities when working with the Python implementation. During conversion from Python Protobuf to C++ Protobuf messages, missing ```required``` fields were not reported from the ```SerializeToString``` call. Internally, ```SerializeToString``` calls ```FindInitializationErrors``` upon failure and raises an error. This fetches that error and raises it as a ```RuntimeError```.

While not as nicely formatted as C++ provided by #94, it is an improvement over the previously generic "RuntimeError: Failed to Serialize python protobuf message.".